requirements-dev.txt
#docs
Sphinx

#linting
flake8

#testing
pytest
python-dotenv
pyyaml
